extract time functions
